iOS8 でのレイアウト乱れに場当たり的対応 タブバーのアイコンをタップすると一部のタブアイコン文字列の表示が変わってしまう見苦しい問題を修正 リスト表示計ビューコントローラーがタブバーの下半分をわずかに黒く変色させていた問題を修正 地図画面の一覧表示を出した状態から駅の乗下車情報を変更した場合、乗下車済／未乗下車のアイコン表示が変わらなかった問題を修正
git-svn-id: http://svn.wsf-lp.com/svn/iOritsubushi@167 dca55b56-fcf5-11e1-90ce-9987a2ddb210

+++ Oritsubushi/PlaceSelectView.m
@@ -39,7 +39,7 @@
         [self addButtonWithTitle:NSLocalizedString(@"キャンセル", nil)];
         self.cancelButtonIndex = 0;
         
-        int numRows = [candidates count];
+        NSInteger numRows = [candidates count];
         CGRect frame = CGRectMake(0,
                                   0,
                                   100,
